Strip class methods from MonomialSum to free functionns
- rename test_gem_optimise.py to test_coffee_optimise.py
