+++ gem/refactorise.py
@@ -2,20 +2,16 @@
 refactorisation."""
 
 from __future__ import absolute_import, print_function, division
-from six import iteritems, itervalues, iterkeys
-from six.moves import intern, map, filter, filterfalse
+from six import iteritems
+from six.moves import intern, map
 
 from collections import Counter, OrderedDict, defaultdict, namedtuple
-from itertools import product, count
-
-import numpy
+from itertools import product
 
 from gem.node import Memoizer, traversal
-from gem.gem import (Node, Zero, Product, Sum, Indexed, ListTensor, one,
-                     IndexSum)
+from gem.gem import (Node, Zero, Product, Sum, Indexed, ListTensor, one)
 from gem.optimise import (remove_componenttensors, sum_factorise,
-                          traverse_product, traverse_sum, unroll_indexsum,
-                          fast_sum_factorise, associate_product, associate_sum)
+                          traverse_product, traverse_sum, unroll_indexsum)
 
 
 # Refactorisation labels
@@ -115,176 +111,6 @@ class MonomialSum(object):
                 rest = Product(r, rest)
             result.add(sum_indices, atomics, rest)
         return result
-
-    def argument_indices_extent(self, factor):
-        """
-        Returns the product of extents of argument indices of :param: factor
-        """
-        if self.argument_indices is None:
-            raise AssertionError("argument_indices property not initialised.")
-        return numpy.product([i.extent for i in set(factor.free_indices).intersection(self.argument_indices)])
-
-    def unique_sum_indices(self):
-        """generator of unique sum indices, together with their original
-        ordering of this MonomialSum
-        """
-        seen = set()
-        for (sum_indices_set, _), (sum_indices, _) in iteritems(self.ordering):
-            if sum_indices_set not in seen:
-                seen.add(sum_indices_set)
-                yield (sum_indices_set, sum_indices)
-
-    def to_expression(self):
-        """
-        Convert MonomialSum object to gem node. Use associate_product() and
-        associate_sum() to promote hoisting in subsequent code generation.
-        ordering ensures deterministic code generation.
-        :return: gem node represented by this MonomialSum object.
-        """
-        indexsums = []  # The result is summation of indexsums
-        monomial_group = OrderedDict()  # (sum_indices_set, sum_indices) -> [(atomics, rest)]
-        # Group monomials according to their sum indices
-        for key, (sum_indices, atomics) in iteritems(self.ordering):
-            sum_indices_set, _ = key
-            rest = self.monomials[key]
-            if not sum_indices_set:
-                indexsums.append(fast_sum_factorise(sum_indices, atomics + (rest,)))
-            else:
-                monomial_group.setdefault((sum_indices_set, sum_indices), []).append((atomics, rest))
-
-        # Form IndexSum's from each monomial group
-        for (_, sum_indices), monomials in iteritems(monomial_group):
-            all_atomics, all_rest = zip(*monomials)
-            if len(all_atomics) == 1:
-                # Just one term, add to indexsums directly
-                indexsums.append(fast_sum_factorise(sum_indices, all_atomics[0] + (all_rest[0],)))
-            else:
-                # Form products for each monomial
-                products = [associate_product(atomics + (_rest,))[0] for atomics, _rest in zip(all_atomics, all_rest)]
-                indexsums.append(IndexSum(associate_sum(products)[0], sum_indices))
-
-        return associate_sum(indexsums)[0]
-
-    def find_optimal_atomics(self, sum_indices_set):
-        index = count()
-        atomic_index = OrderedDict()  # Atomic gem node -> int
-        connections = []
-        # add connections (list of lists)
-        for (_sum_indices, _), (_, atomics) in iteritems(self.ordering):
-            if _sum_indices == sum_indices_set:
-                connection = []
-                for atomic in atomics:
-                    if atomic not in atomic_index:
-                        atomic_index[atomic] = next(index)
-                    connection.append(atomic_index[atomic])
-                connections.append(tuple(connection))
-
-        if len(atomic_index) == 0:
-            return ((), ())
-        if len(atomic_index) == 1:
-            return ((list(atomic_index.keys())[0], ), ())
-
-        # set up the ILP
-        import pulp as ilp
-        ilp_prob = ilp.LpProblem('gem factorise', ilp.LpMinimize)
-        ilp_var = ilp.LpVariable.dicts('node', range(len(atomic_index)), 0, 1, ilp.LpBinary)
-
-        # Objective function
-        # Minimise number of factors to pull. If same number, favour factor with larger extent
-        big = 10000000  # some arbitrary big number
-        ilp_prob += ilp.lpSum(ilp_var[index] * (big - self.argument_indices_extent(atomic)) for atomic, index in iteritems(atomic_index))
-
-        # constraints
-        for connection in connections:
-            ilp_prob += ilp.lpSum(ilp_var[index] for index in connection) >= 1
-
-        ilp_prob.solve()
-        if ilp_prob.status != 1:
-            raise AssertionError("Something bad happened during ILP")
-
-        def optimal(atomic):
-            return ilp_var[atomic_index[atomic]].value() == 1
-
-        optimal_atomics = filter(optimal, iterkeys(atomic_index))
-        other_atomics = filterfalse(optimal, iterkeys(atomic_index))
-
-        return (tuple(optimal_atomics), tuple(other_atomics))
-
-    def factorise_atomics(self, optimal_atomics):
-        """
-        Group and factorise monomials based on a list of atomics. Create new monomials for
-        each group and optimise them recursively.
-        :param optimal_atomics: list of tuples of optimal atomics and their sum indices
-        :return: new MonomialSum object with atomics factorised, or self if no changes are made
-        """
-        if not optimal_atomics:
-            return self
-        if len(self.ordering) < 2:
-            return self
-        new_monomial_sum = MonomialSum()
-        new_monomial_sum.argument_indices = self.argument_indices
-        # Group monomials according to each optimal atomic
-        factor_group = OrderedDict()
-        for key, (_sum_indices, _atomics) in iteritems(self.ordering):
-            for (sum_indices_set, sum_indices), oa in optimal_atomics:
-                if key[0] == sum_indices_set and oa in _atomics:
-                    # Add monomial key to the list of corresponding optimal atomic
-                    factor_group.setdefault(((sum_indices_set, sum_indices), oa), []).append(key)
-                    break
-            else:
-                # Add monomials that do no have argument factors to new MonomialSum
-                new_monomial_sum.add(_sum_indices, _atomics, self.monomials[key])
-        # We should not drop monomials
-        assert sum(map(len, itervalues(factor_group))) + len(new_monomial_sum.ordering) == len(self.ordering)
-
-        for ((sum_indices_set, sum_indices), oa), keys in iteritems(factor_group):
-            if len(keys) == 1:
-                # Just one monomials with this atomic, add to new MonomialSum straightaway
-                _, _atomics = self.ordering[keys[0]]
-                _rest = self.monomials[keys[0]]
-                new_monomial_sum.add(sum_indices, _atomics, _rest)
-                continue
-            all_atomics = []  # collect all atomics from monomials
-            all_rest = []  # collect all rest from monomials
-            for key in keys:
-                _, _atomics = self.ordering[key]
-                _atomics = list(_atomics)
-                _atomics.remove(oa)  # remove common factor
-                all_atomics.append(_atomics)
-                all_rest.append(self.monomials[key])
-            # Create new MonomialSum for the factorised out terms
-            sub_monomial_sum = MonomialSum()
-            sub_monomial_sum.argument_indices = self.argument_indices
-            for _atomics, _rest in zip(all_atomics, all_rest):
-                sub_monomial_sum.add((), _atomics, _rest)
-            sub_monomial_sum = sub_monomial_sum.optimise()
-            assert len(sub_monomial_sum.ordering) != 0
-            if len(sub_monomial_sum.ordering) == 1:
-                # result is a product, add to new MonomialSum directly
-                (_, new_atomics), = itervalues(sub_monomial_sum.ordering)
-                new_atomics += (oa,)
-                new_rest, = itervalues(sub_monomial_sum.monomials)
-            else:
-                # result is a sum, need to form new node
-                new_node = sub_monomial_sum.to_expression()
-                new_atomics = [oa]
-                new_rest = one
-                if set(self.argument_indices) & set(new_node.free_indices):
-                    new_atomics.append(new_node)
-                else:
-                    new_rest = new_node
-            new_monomial_sum.add(sum_indices, new_atomics, new_rest)
-        return new_monomial_sum
-
-    def optimise(self):
-        all_optimal_atomics = []  # [((sum_indices_set, sum_indces), optimal_atomics))]
-        for sum_indices_set, sum_indices in self.unique_sum_indices():
-            # throw away other atomics here
-            optimal_atomics, _ = self.find_optimal_atomics(sum_indices_set)
-            all_optimal_atomics.extend([((sum_indices_set, sum_indices), _atomic) for _atomic in optimal_atomics])
-        # This algorithm is O(N!), where N = len(optimal_atomics)
-        # we could truncate the optimal_atomics list at say 10
-        return self.factorise_atomics(all_optimal_atomics)
 
 
 class FactorisationError(Exception):
