+++ tests/test_coffee_optimise.py
@@ -3,7 +3,9 @@ from __future__ import absolute_import, print_function, division
 import pytest
 
 from gem.gem import Index, Indexed, Product, Variable, Division, Literal, Sum
-from gem.optimise import replace_division, reassociate_product, optimise, count_flop
+from gem.optimise import replace_division, reassociate_product
+from gem.utils import count_flop
+from tsfc.coffee_mode import optimise
 
 
 def test_replace_div():
